+++ examples/react-typescript-cs16-webrtc/src/App.tsx
@@ -4,6 +4,7 @@ import xashURL from 'xash3d-fwgs/xash.wasm'
 import menuURL from 'cs16-client/cl_dll/menu_emscripten_wasm32.wasm'
 import clientURL from 'cs16-client/cl_dll/client_emscripten_wasm32.wasm'
 import serverURL from 'cs16-client/dlls/cs_emscripten_wasm32.so'
+import extrasURL from 'cs16-client/extras.pk3'
 import gles3URL from 'xash3d-fwgs/libref_gles3compat.wasm'
 import {loadAsync} from 'jszip'
 import {Xash3DWebRTC} from "./webrtc";
@@ -17,9 +18,7 @@ const App: FC = () => {
             <button className="Input" onClick={async () => {
                 const x = new Xash3DWebRTC({
                     canvas: canvasRef.current!,
-                    module: {
-                        arguments: ['-windowed', '-game', 'cstrike'],
-                    },
+                    arguments: ['-windowed', '-game', 'cstrike'],
                     libraries: {
                         filesystem: filesystemURL,
                         xash: xashURL,
@@ -30,16 +29,21 @@ const App: FC = () => {
                             gles3compat: gles3URL,
                         }
                     },
+                    dynamicLibraries: ['dlls/cs_emscripten_wasm32.so', '/rwdir/filesystem_stdio.wasm'],
                     filesMap: {
                         'dlls/cs_emscripten_wasm32.so': serverURL,
-                        '/rwdir/filesystem_stdio.so': filesystemURL,
+                        '/rwdir/filesystem_stdio.wasm': filesystemURL,
                     },
                 });
 
-                const [zip] = await Promise.all([
+                const [zip, extras] = await Promise.all([
                     (async () => {
                         const res = await fetch('valve.zip')
                         return await loadAsync(await res.arrayBuffer());
+                    })(),
+                    (async () => {
+                        const res = await fetch(extrasURL)
+                        return await res.arrayBuffer()
                     })(),
                     x.init(),
                 ])
@@ -55,6 +59,8 @@ const App: FC = () => {
                     x.em.FS.mkdirTree(dir);
                     x.em.FS.writeFile(path, await file.async("uint8array"));
                 }))
+
+                x.em.FS.writeFile('/rodir/cstrike/extras.pk3', new Uint8Array(extras))
 
                 x.em.FS.chdir('/rodir')
                 x.main()