+++ examples/react-typescript-hlsdk-webrtc/src/App.tsx
@@ -5,6 +5,7 @@ import menuURL from 'xash3d-fwgs/cl_dll/menu_emscripten_wasm32.wasm'
 import clientURL from 'hlsdk-portable/cl_dll/client_emscripten_wasm32.wasm'
 import serverURL from 'hlsdk-portable/dlls/hl_emscripten_wasm32.so'
 import gles3URL from 'xash3d-fwgs/libref_gles3compat.wasm'
+import extrasURL from 'xash3d-fwgs/extras.pk3'
 import {loadAsync} from 'jszip'
 import {Xash3DWebRTC} from "./webrtc";
 import './App.css';
@@ -17,9 +18,7 @@ const App: FC = () => {
             <button className="Input" onClick={async () => {
                 const x = new Xash3DWebRTC({
                     canvas: canvasRef.current!,
-                    module: {
-                        arguments: ['-windowed'],
-                    },
+                    arguments: ['-windowed'],
                     libraries: {
                         filesystem: filesystemURL,
                         xash: xashURL,
@@ -32,10 +31,14 @@ const App: FC = () => {
                     }
                 });
 
-                const [zip] = await Promise.all([
+                const [zip, extras] = await Promise.all([
                     (async () => {
                         const res = await fetch('valve.zip')
                         return await loadAsync(await res.arrayBuffer());
+                    })(),
+                    (async () => {
+                        const res = await fetch(extrasURL)
+                        return await res.arrayBuffer()
                     })(),
                     x.init(),
                 ])
@@ -51,6 +54,8 @@ const App: FC = () => {
                     x.em.FS.mkdirTree(dir);
                     x.em.FS.writeFile(path, await file.async("uint8array"));
                 }))
+
+                x.em.FS.writeFile('/rodir/valve/extras.pk3', new Uint8Array(extras))
 
                 x.em.FS.chdir('/rodir')
                 x.main()