+++ packages/xash3d-fwgs/lib/net/index.ts
@@ -1,0 +1,2 @@
+export * from './defaultNet'
+export * from './net'