+++ packages/xash3d-fwgs/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-09-03
+### Changed
+- gles3compat renderer wasm replaced with gl4es and will be removed in the next major release
+- Extended net module
+- Added engine callbacks
+
+---
+
 ## [1.0.1] - 2025-08-01
 ### Fixed
 - Screen lock without Net